+++ CHANGELOG.md
@@ -13,6 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Security` in case of vulnerabilities.
 
 
+## [0.8.2] - Unreleased
+
+### Added
+- console more info for unmatched sandbox
+- `docit.config.js` now supports `vite` options
+
+### Fixes
+- remove outer div of sandbox
+
+### Changed
+- changed toc and sidebar style
+- adjust mobile view sandbox code height to 640px
+
+
 ## [0.8.1] - 2022-04-18
 
 ### Fixed